Email verification code formatting.
